Fix overflow bug

+++ String/tests/lcp.test.cpp
@@ -11,7 +11,7 @@ using namespace std;
 int32_t main() {
     ios::sync_with_stdio(0); cin.tie(0);
     string s; cin >> s;
-    int res = SZ(s) * (SZ(s) + 1) / 2;
+    long long res = SZ(s) * (long long)(SZ(s) + 1) / 2;
     SuffixArray sa(s);
     for (int x : sa.LCP) {
         res -= x;
